% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_end_correction.R
\name{annotatepolyAsignals}
\alias{annotatepolyAsignals}
\title{Title}
\usage{
annotatepolyAsignals(gr.object, APAmotifs, hexamers)
}
\arguments{
\item{gr.object}{}

\item{APAmotifs}{}

\item{hexamers}{}
}
\description{
Title
}
