% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_end_correction.R
\name{blockGuidedcorrection}
\alias{blockGuidedcorrection}
\title{Title}
\usage{
blockGuidedcorrection(blocks, isoform.assembly, LongWindow, shortWindow)
}
\arguments{
\item{blocks}{}

\item{isoform.assembly}{}

\item{LongWindow}{}

\item{shortWindow}{}
}
\description{
Title
}
