% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_read_processing.R
\name{create_clusters}
\alias{create_clusters}
\title{Create 3'end clusters}
\usage{
create_clusters(nucleotide_regions, refAnnot, window)
}
\arguments{
\item{window}{}
}
\description{
Create 3'end clusters
}
