% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_end_correction.R
\name{createScaffoldBlocks}
\alias{createScaffoldBlocks}
\title{Title}
\usage{
createScaffoldBlocks(isoforms.sqanti, LongEnds)
}
\arguments{
\item{isoforms.sqanti}{}

\item{LongEnds}{}
}
\description{
Title
}
