% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_read_processing.R
\name{extract_features}
\alias{extract_features}
\title{Extract features from clusters}
\usage{
extract_features(clusters3seq, genomeSeq, dwindow, upwindow)
}
\arguments{
\item{upwindow}{}
}
\description{
Extract features from clusters
}
