% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_feature_extraction.R
\name{getIsoformFeatures}
\alias{getIsoformFeatures}
\title{prepare data for 5'-3' window filtering
Create a reference database of all TSS-TES link associations based on annotation}
\usage{
getIsoformFeatures(annotation, promoter_window, pas_window)
}
\arguments{
\item{annotation}{reference annotation GenomicRanges object}

\item{promoter_window}{window to consider distinct TSS}

\item{pas_window}{window to consider distinct TES}
}
\value{
promoterDatabase with classification of links
}
\description{
prepare data for 5'-3' window filtering
Create a reference database of all TSS-TES link associations based on annotation
}
