% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_end_correction.R
\name{getLastExonperTranscript}
\alias{getLastExonperTranscript}
\title{Title}
\usage{
getLastExonperTranscript(seq.gr)

getLastExonperTranscript(seq.gr)
}
\arguments{
\item{seq.gr}{}
}
\description{
Title

Title
}
