% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_end_correction.R,
%   R/long_read_processing.R
\name{getPolyAsignal}
\alias{getPolyAsignal}
\title{Title}
\usage{
getPolyAsignal(APAmotifs, hexamers, seqs)

getPolyAsignal(APAmotifs, hexamers, seqs)
}
\arguments{
\item{APAmotifs}{}

\item{hexamers}{}

\item{seqs}{}
}
\description{
Title

Title
}
