% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_end_correction.R
\name{gRanges2Biostrings3UTR}
\alias{gRanges2Biostrings3UTR}
\title{Title}
\usage{
gRanges2Biostrings3UTR(grset)

gRanges2Biostrings3UTR(grset)
}
\arguments{
\item{grset}{}
}
\description{
Title

Title
}
