% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_end_correction.R
\name{strandAwareFindOverlapEnd}
\alias{strandAwareFindOverlapEnd}
\title{Title}
\usage{
strandAwareFindOverlapEnd(queryData, annotation, maxgap)
}
\arguments{
\item{queryData}{}

\item{annotation}{}

\item{maxgap}{}
}
\description{
Title
}
