% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_read_processing.R
\name{trim_single_nucleotide}
\alias{trim_single_nucleotide}
\title{Trim long reads to single nucleotide in their 3'ends}
\usage{
trim_single_nucleotide(reads.gr)
}
\arguments{
\item{GenomicRanges}{object from GenomicAlignments::readGAlignments() bam file}
}
\value{
trimmed genomic ranges object
}
\description{
Trim long reads to single nucleotide in their 3'ends
}
